requirements.txt
# requirements.txt

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
mysql-connector-python==8.2.0

# Security & Authentication
argon2-cffi==23.1.0
python-dotenv==1.0.0

# Image Processing
Pillow==10.1.0

# Utilities
email-validator==2.1.0